+++ CapstoneProfessions/Framework/CapstoneProfessionMenu.cs
@@ -81,7 +81,7 @@ namespace CapstoneProfessions.Framework
             this.extraInfoForLevel.Clear();
             Game1.player.completelyStopAnimatingOrDoingAction();
             this.informationUp = true;
-            this.title = Mod.Instance.Helper.Translation.Get("menu.title");
+            this.title = I18n.Menu_Title();
             this.extraInfoForLevel = this.getExtraInfoForLevel();
             this.sourceRectForLevelIcon = Game1.whichFarm switch
             {
@@ -160,7 +160,7 @@ namespace CapstoneProfessions.Framework
         {
             return new()
             {
-                Mod.Instance.Helper.Translation.Get("menu.extra")
+                I18n.Menu_Extra()
             };
         }
 
@@ -183,8 +183,8 @@ namespace CapstoneProfessions.Framework
             {
                 this.professionsToChoose.Add(Mod.ProfessionTime);
                 this.professionsToChoose.Add(Mod.ProfessionProfit);
-                this.leftProfessionDescription = new List<string>(new string[] { Mod.Instance.Helper.Translation.Get("profession.time.name"), Mod.Instance.Helper.Translation.Get("profession.time.description") });
-                this.rightProfessionDescription = new List<string>(new string[] { Mod.Instance.Helper.Translation.Get("profession.profit.name"), Mod.Instance.Helper.Translation.Get("profession.profit.description") });
+                this.leftProfessionDescription = new List<string>(new[] { I18n.Profession_Time_Name(), I18n.Profession_Time_Description() });
+                this.rightProfessionDescription = new List<string>(new[] { I18n.Profession_Profit_Name(), I18n.Profession_Profit_Description() });
                 this.hasUpdatedProfessions = true;
             }
             for (int i = this.littleStars.Count - 1; i >= 0; i--)
@@ -317,7 +317,7 @@ namespace CapstoneProfessions.Framework
                     Utility.drawWithShadow(b, Game1.mouseCursors, new Vector2(this.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth, this.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + 16), this.sourceRectForLevelIcon, Color.White, 0f, Vector2.Zero, 4f, flipped: false, 0.88f);
                     b.DrawString(Game1.dialogueFont, this.title, new Vector2(this.xPositionOnScreen + this.width / 2 - Game1.dialogueFont.MeasureString(this.title).X / 2f, this.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + 16), Game1.textColor);
                     Utility.drawWithShadow(b, Game1.mouseCursors, new Vector2(this.xPositionOnScreen + this.width - IClickableMenu.spaceToClearSideBorder - IClickableMenu.borderWidth - 64, this.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + 16), this.sourceRectForLevelIcon, Color.White, 0f, Vector2.Zero, 4f, flipped: false, 0.88f);
-                    string chooseProfession = Mod.Instance.Helper.Translation.Get("menu.extra");
+                    string chooseProfession = I18n.Menu_Extra();
                     b.DrawString(Game1.smallFont, chooseProfession, new Vector2(this.xPositionOnScreen + this.width / 2 - Game1.smallFont.MeasureString(chooseProfession).X / 2f, this.yPositionOnScreen + 64 + IClickableMenu.spaceToClearTopBorder), Game1.textColor);
                     b.DrawString(Game1.dialogueFont, this.leftProfessionDescription[0], new Vector2(this.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + 32, this.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + 160), this.leftProfessionColor);
                     b.Draw(Mod.ClockTex, new Vector2(this.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + this.width / 2 - 112, this.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + 160 - 16), null, Color.White, 0f, Vector2.Zero, 2f, SpriteEffects.None, 1f);