+++ PreexistingRelationship/Framework/MarryMenu.cs
@@ -47,7 +47,7 @@ namespace PreexistingRelationship.Framework
 
             var title = new Label
             {
-                String = Mod.Instance.Helper.Translation.Get("menu.title"),
+                String = I18n.Menu_Title(),
                 Bold = true
             };
             title.LocalPosition = new Vector2((800 - title.Measure().X) / 2, 10);
@@ -55,7 +55,7 @@ namespace PreexistingRelationship.Framework
 
             this.Ui.AddChild(new Label
             {
-                String = Mod.Instance.Helper.Translation.Get("menu.text").ToString().Replace("\\n", "\n"),
+                String = I18n.Menu_Text(),
                 LocalPosition = new Vector2(50, 75),
                 NonBoldScale = 0.75f,
                 NonBoldShadow = false
@@ -126,13 +126,13 @@ namespace PreexistingRelationship.Framework
 
             this.Ui.AddChild(new Label
             {
-                String = Mod.Instance.Helper.Translation.Get("menu.button.cancel"),
+                String = I18n.Menu_Button_Cancel(),
                 LocalPosition = new Vector2(175, 650),
                 Callback = e => Game1.exitActiveMenu()
             });
             this.Ui.AddChild(new Label
             {
-                String = Mod.Instance.Helper.Translation.Get("menu.button.accept"),
+                String = I18n.Menu_Button_Accept(),
                 LocalPosition = new Vector2(500, 650),
                 Callback = e => this.DoMarriage()
             });
@@ -168,7 +168,7 @@ namespace PreexistingRelationship.Framework
             {
                 if (player.spouse == this.SelectedNpc)
                 {
-                    Game1.addHUDMessage(new HUDMessage(Mod.Instance.Helper.Translation.Get("spouse-taken")));
+                    Game1.addHUDMessage(new HUDMessage(I18n.SpouseTaken()));
                     this.SelectedContainer.OutlineColor = null;
                     this.SelectedContainer = null;
                     this.SelectedNpc = null;